Completed exercises 4.15-4.23

+++ part4/blog/controllers/blogs.js
@@ -1,26 +1,30 @@
 const blogsRouter = require("express").Router()
 const Blog = require("../models/blog")
+const User = require("../models/user")
+const jwt = require("jsonwebtoken")
 
-blogsRouter.get("/", (request, response) => {
-  Blog.find({}).then((blogs) => {
-    response.json(blogs)
+blogsRouter.get("/", async (request, response) => {
+  const blogs = await Blog.find({}).populate("user", {
+    name: 1,
+    id: 1,
+    username: 1,
   })
+  response.json(blogs)
 })
 
-blogsRouter.get("/:id", (request, response, next) => {
-  Blog.findById(request.params.id)
-    .then((blog) => {
-      if (blog) {
-        response.json(blog)
-      } else {
-        response.status(404).end()
-      }
-    })
-    .catch((error) => next(error))
+blogsRouter.get("/:id", async (request, response, next) => {
+  const blog = await Blog.findById(request.params.id)
+  response.json(blog)
 })
 
-blogsRouter.post("/", (request, response, next) => {
+blogsRouter.post("/", async (request, response, next) => {
   const body = request.body
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" })
+  }
+  const user = request.user
 
   if (!body.title || !body.url) {
     return response.status(400).end()
@@ -30,27 +34,39 @@ blogsRouter.post("/", (request, response, next) => {
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
+    user: user.id,
   })
 
-  blog
-    .save()
-    .then((savedBlog) => {
-      response.json(savedBlog)
-    })
-    .catch((error) => next(error))
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+  response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete("/:id", (request, response, next) => {
-  Blog.findByIdAndDelete(request.params.id)
-    .then(() => {
-      response.status(204).end()
-    })
-    .catch((error) => next(error))
+blogsRouter.delete("/:id", async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" })
+  }
+  const user = request.user
+  if (user.id.toString() !== decodedToken.id.toString()) {
+    return response.status(403).json({ error: "permission denied" })
+  }
+  await Blog.findByIdAndDelete(request.params.id)
+  response.status(204).end()
 })
 
-blogsRouter.put("/:id", (request, response, next) => {
+blogsRouter.put("/:id", async (request, response, next) => {
   const body = request.body
 
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" })
+  }
+  const user = request.user
+  if (user.id.toString() !== decodedToken.id.toString()) {
+    return response.status(403).json({ error: "permission denied" })
+  }
   const blog = {
     title: body.title,
     author: body.author,
@@ -58,11 +74,10 @@ blogsRouter.put("/:id", (request, response, next) => {
     likes: body.likes,
   }
 
-  Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
-    .then((updatedBlog) => {
-      response.json(updatedBlog)
-    })
-    .catch((error) => next(error))
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {
+    new: true,
+  })
+  response.status(201).json(updatedBlog)
 })
 
 module.exports = blogsRouter
